Fixed bug in lead details
